rename to `inlayHintsProvider` to match conventions
