+++ src/extension.ts
@@ -10,7 +10,7 @@ import {
 } from 'vscode-languageclient/node';
 import { EffektManager } from './effektManager';
 import { EffektIRContentProvider } from './irProvider';
-import { InlayHintProvider } from './InlayHintsProvider';
+import { InlayHintProvider } from './inlayHintsProvider';
 import * as net from 'net';
 
 /**
@@ -286,7 +286,7 @@ export async function activate(context: vscode.ExtensionContext) {
         { scheme: 'file', language: 'effekt' },
         new InlayHintProvider(client)
     );
-    
+
     vscode.languages.registerInlayHintsProvider(
         { scheme: 'file', language: 'literate effekt' },
         new InlayHintProvider(client)
